[BAEL-8232] - Moved java string related code and github links into new module java-strings
