+++ core-java/README.md
@@ -3,18 +3,14 @@
 ## Core Java Cookbooks and Examples
 
 ### Relevant Articles: 
-- [Java – Generate Random String](http://www.baeldung.com/java-random-string)
 - [Java Timer](http://www.baeldung.com/java-timer-and-timertask)
 - [How to Run a Shell Command in Java](http://www.baeldung.com/run-shell-command-in-java)
 - [MD5 Hashing in Java](http://www.baeldung.com/java-md5)
 - [Guide to Java Reflection](http://www.baeldung.com/java-reflection)
 - [A Guide to Java Sockets](http://www.baeldung.com/a-guide-to-java-sockets)
-- [Convert char to String in Java](http://www.baeldung.com/java-convert-char-to-string)
-- [Convert String to int or Integer in Java](http://www.baeldung.com/java-convert-string-to-int-or-integer)
 - [Java – Try with Resources](http://www.baeldung.com/java-try-with-resources)
 - [Guide to the Fork/Join Framework in Java](http://www.baeldung.com/java-fork-join)
 - [How to Print Screen in Java](http://www.baeldung.com/print-screen-in-java)
-- [How to Convert String to different data types in Java](http://www.baeldung.com/java-string-conversions)
 - [Introduction to Java Generics](http://www.baeldung.com/java-generics)
 - [Generate equals() and hashCode() with Eclipse](http://www.baeldung.com/java-eclipse-equals-and-hashcode)
 - [A Guide To Java Regular Expressions API](http://www.baeldung.com/regular-expressions-java)
@@ -37,8 +33,6 @@
 - [A Quick JUnit vs TestNG Comparison](http://www.baeldung.com/junit-vs-testng)
 - [Java Primitive Conversions](http://www.baeldung.com/java-primitive-conversions)
 - [Java Money and the Currency API](http://www.baeldung.com/java-money-and-currency)
-- [Converting Strings to Enums in Java](http://www.baeldung.com/java-string-to-enum)
-- [Quick Guide to the Java StringTokenizer](http://www.baeldung.com/java-stringtokenizer)
 - [JVM Log Forging](http://www.baeldung.com/jvm-log-forging)
 - [Guide to sun.misc.Unsafe](http://www.baeldung.com/java-unsafe)
 - [How to Perform a Simple HTTP Request in Java](http://www.baeldung.com/java-http-request)
@@ -55,28 +49,22 @@
 - [How to Copy an Array in Java](http://www.baeldung.com/java-array-copy)
 - [Period and Duration in Java](http://www.baeldung.com/java-period-duration)
 - [Converting a Stack Trace to a String in Java](http://www.baeldung.com/java-stacktrace-to-string)
-- [Count Occurrences of a Char in a String](http://www.baeldung.com/java-count-chars)
 - [Java Double Brace Initialization](http://www.baeldung.com/java-double-brace-initialization)
 - [The StackOverflowError in Java](http://www.baeldung.com/java-stack-overflow-error)
-- [Split a String in Java](http://www.baeldung.com/java-split-string)
 - [Introduction to Java Serialization](http://www.baeldung.com/java-serialization)
-- [How to Remove the Last Character of a String?](http://www.baeldung.com/java-remove-last-character-of-string)
 - [ClassNotFoundException vs NoClassDefFoundError](http://www.baeldung.com/java-classnotfoundexception-and-noclassdeffounderror)
 - [Guide to UUID in Java](http://www.baeldung.com/java-uuid)
 - [Guide to Escaping Characters in Java RegExps](http://www.baeldung.com/java-regexp-escape-char)
 - [Guide to hashCode() in Java](http://www.baeldung.com/java-hashcode)
 - [Difference between URL and URI](http://www.baeldung.com/java-url-vs-uri)
 - [Broadcasting and Multicasting in Java](http://www.baeldung.com/java-broadcast-multicast)
-- [CharSequence vs. String in Java](http://www.baeldung.com/java-char-sequence-string)
 - [Period and Duration in Java](http://www.baeldung.com/java-period-duration)
 - [Guide to the Diamond Operator in Java](http://www.baeldung.com/java-diamond-operator)
 - [“Sneaky Throws” in Java](http://www.baeldung.com/java-sneaky-throws)
 - [OutOfMemoryError: GC Overhead Limit Exceeded](http://www.baeldung.com/java-gc-overhead-limit-exceeded)
-- [StringBuilder and StringBuffer in Java](http://www.baeldung.com/java-string-builder-string-buffer)
 - [Creating a Java Compiler Plugin](http://www.baeldung.com/java-build-compiler-plugin)
 - [A Guide to the Static Keyword in Java](http://www.baeldung.com/java-static)
 - [Initializing Arrays in Java](http://www.baeldung.com/java-initialize-array)
-- [Guide to Java String Pool](http://www.baeldung.com/java-string-pool)
 - [Quick Example - Comparator vs Comparable in Java](http://www.baeldung.com/java-comparator-comparable)
 - [Quick Guide to Java Stack](http://www.baeldung.com/java-stack)
 - [The Java continue and break Keywords](http://www.baeldung.com/java-continue-and-break)
@@ -102,8 +90,6 @@
 - [The Trie Data Structure in Java](http://www.baeldung.com/trie-java)
 - [Introduction to Javadoc](http://www.baeldung.com/javadoc)
 - [How to Make a Deep Copy of an Object in Java](http://www.baeldung.com/java-deep-copy)
-- [Check if a String is a Palindrome](http://www.baeldung.com/java-palindrome)
-- [Comparing Strings in Java](http://www.baeldung.com/java-compare-strings)
 - [Guide to Inheritance in Java](http://www.baeldung.com/java-inheritance)
 - [Guide to Externalizable Interface in Java](http://www.baeldung.com/java-externalizable)
 - [Object Type Casting in Java](http://www.baeldung.com/java-type-casting)
@@ -135,11 +121,9 @@
 - [Infinite Loops in Java](http://www.baeldung.com/infinite-loops-java)
 - [Using Java Assertions](http://www.baeldung.com/java-assert)
 - [Pass-By-Value as a Parameter Passing Mechanism in Java](http://www.baeldung.com/java-pass-by-value-or-pass-by-reference)
-- [Check If a String Is Numeric in Java](http://www.baeldung.com/java-check-string-number)
 - [Variable and Method Hiding in Java](http://www.baeldung.com/java-variable-method-hiding)
 - [Access Modifiers in Java](http://www.baeldung.com/java-access-modifiers)
 - [Infinite Loops in Java](http://www.baeldung.com/infinite-loops-java)
-- [Why Use char[] Array Over a String for Storing Passwords in Java?](http://www.baeldung.com/java-storing-passwords)
 - [Introduction to Creational Design Patterns](http://www.baeldung.com/creational-design-patterns)
 - [Proxy, Decorator, Adapter and Bridge Patterns](http://www.baeldung.com/java-structural-design-patterns)
 - [Singletons in Java](http://www.baeldung.com/java-singleton)
@@ -153,7 +137,6 @@
 - [Importance of Main Manifest Attribute in a Self-Executing JAR](http://www.baeldung.com/java-jar-executable-manifest-main-class)
 - [Extracting Year, Month and Day from Date in Java](http://www.baeldung.com/java-year-month-day)
 - [Get Date Without Time in Java](http://www.baeldung.com/java-date-without-time)
-- [Convert a String to Title Case](http://www.baeldung.com/java-string-title-case)
 - [How to Get the File Extension of a File in Java](http://www.baeldung.com/java-file-extension)
 - [Immutable Objects in Java](http://www.baeldung.com/java-immutable-object)
 - [Console I/O in Java](http://www.baeldung.com/java-console-input-output)
