+++ java-strings/src/main/java/com/baeldung/java9/compactstring/CompactStringDemo.java
