+++ java-strings/src/main/java/com/baeldung/string/Palindrome.java
