+++ java-strings/src/main/resources/data.csv
