+++ java-strings/src/test/java/com/baeldung/CharArrayToStringUnitTest.java
