+++ java-strings/src/test/java/com/baeldung/chararraypassword/PasswordStoreExamplesUnitTest.java
