+++ java-strings/src/test/java/com/baeldung/string/StringToCharStreamUnitTest.java
