+++ java-strings/src/test/java/com/baeldung/string/formatter/StringFormatterExampleUnitTest.java
