+++ java-strings/src/test/java/com/baeldung/stringisnumeric/CoreJavaIsNumericUnitTest.java
